Refactor: 자주 쓰이는 타입을 전역 타입으로 지정 및 바르지 않은 파일 이름 정정
