+++ src/modal/AddLinkModal.tsx
@@ -1,16 +1,11 @@
 import ModalLayout from "./ModalLayout";
 import { useState } from "react";
-import { MODAL_TYPE } from "./modalType";
+import { MODAL_TYPE } from "../constants/modalConstans";
 import { DATA } from "../constants/folderlist_mock_data";
 import IconCheck from "../images/Icon_check.svg";
 import styles from "./AddLinkModal.module.css";
 
-interface Props {
-  isOpenModal: boolean,
-  closeModal: () => void,
-}
-
-function AddLinkModal({ isOpenModal, closeModal }: Props) {
+function AddLinkModal({ isOpenModal, closeModal }: ModalBaseProps) {
   const [clickItem, setClickItem] = useState('');
   const { addLink } = MODAL_TYPE;
 
@@ -21,7 +16,7 @@ function AddLinkModal({ isOpenModal, closeModal }: Props) {
   return (
     <ModalLayout
       title={addLink.title}
-      isOpen={isOpenModal}
+      isOpenModal={isOpenModal}
       closeModal={closeModal}
     >
       <div className={styles.linkName}>링크 주소</div>
