+++ src/modal/DeleteModal.tsx
@@ -1,20 +1,15 @@
-import { MODAL_TYPE } from "./modalType";
+import { MODAL_TYPE } from "../constants/modalConstans";
 import ModalLayout from "./ModalLayout";
 import styles from "./DeleteModal.module.css";
 
-interface Props {
-  isOpenModal: boolean,
-  closeModal: () => void,
-}
-
-function DeleteModal({ isOpenModal, closeModal }: Props) {
+function DeleteModal({ isOpenModal, closeModal }: ModalBaseProps) {
 
   const { deleteFolder } = MODAL_TYPE;
 
   return (
     <ModalLayout
       title={deleteFolder.title}
-      isOpen={isOpenModal}
+      isOpenModal={isOpenModal}
       closeModal={closeModal}
     >
       <div className={styles.folderName}>폴더명</div>
