+++ src/modal/EditModal.tsx
@@ -1,21 +1,19 @@
 import ModalLayout from "./ModalLayout";
-import { MODAL_TYPE } from "./modalType";
+import { MODAL_TYPE } from "../constants/modalConstans";
 import styles from "./EditModal.module.css";
 
-interface Props {
-  isOpenModal: boolean,
-  closeModal: () => void,
+interface EditModalProps extends ModalBaseProps {
   placeholder: string,
 }
 
-function EditModal({ isOpenModal, closeModal, placeholder }: Props) {
+function EditModal({ isOpenModal, closeModal, placeholder }: EditModalProps) {
 
   const { edit } = MODAL_TYPE;
 
   return (
     <ModalLayout
       title={edit.title}
-      isOpen={isOpenModal}
+      isOpenModal={isOpenModal}
       closeModal={closeModal}
     >
       <input className={styles.inputText} type="text" placeholder={placeholder} />
