+++ src/modal/ModalLayout.tsx
@@ -2,21 +2,19 @@ import styles from "./ModalLayout.module.css";
 import {ReactComponent as CloseIcon} from '../images/Icon_close.svg';
 import { ReactNode } from "react";
 
-interface Props {
+interface LayoutProps extends ModalBaseProps {
   title: string,
-  isOpen: boolean,
   children: ReactNode,
-  closeModal: () => void,
 }
 
-function ModalLayout({ title, isOpen, children, closeModal }: Props) {
+function ModalLayout({ title, children, isOpenModal, closeModal }: LayoutProps) {
   const closeBtnBackground = 'var(--gray10)';
   const closeBtnIcon = 'var(--gray60)';
   
   return (
     <div
       className={styles.background}
-      style={{ display: isOpen ? "block" : "none" }}
+      style={{ display: isOpenModal ? "block" : "none" }}
     >
       <div className={styles.container}>
         <div className={styles.modalTitle}>{title}</div>
