+++ src/react-app-env.d.ts
@@ -2,3 +2,16 @@
 interface Window {
   Kakao: any
 }
+
+interface Data {
+  url: string;
+  id: number;
+  imageSource?: string;
+  createdAt?: string;
+  description?: string;
+}
+
+interface ModalBaseProps {
+  isOpenModal: boolean,
+  closeModal: () => void,
+}
